ROTOSCOPIE: Multithreading
We process a video using multiple threads to make it faster.
The number of threads is equal to the number of procs in the CPU.

We create multiple excerpts (struct Video) which frames start
somewhere in the original video's frames and which frame count
is equal to frame_count / number_of_threads.